# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: J.A.R.V.I.S. API
  version: 0.8.0
  description: >-
    API da fundação e dos incrementos financeiros de despesas, receitas,
    categorias, recorrências manuais, sugestões determinísticas e referências organizacionais de cartões. Criar uma transaction registra no
    organizador J.A.R.V.I.S. uma movimentação que já ocorreu no mundo real. Uma
    recurrence representa somente um compromisso ou expectativa futura e nunca
    cria uma Expense ou Income automaticamente. Nenhuma operação executa
    pagamento, Pix, transferência bancária, compra, recebimento ou movimentação
    de fundos. Os endpoints financeiros só existem quando o modo single-owner
    temporário está explicitamente habilitado.
...
servers:
  - url: http://127.0.0.1:8080
    description: Execução local com a configuração padrão
security: []
paths:
...
  /v1/transactions/preview:
    post:
      operationId: previewTransaction
      summary: Valida uma despesa ou receita para revisão antes da confirmação
      description: >-
        Despacha explicitamente por type e normaliza uma Expense ou Income sem
        gerar ID, gravar transaction, criar audit event ou reservar idempotência.
        O cliente deve apresentar este preview ao usuário antes de chamar a
        operação mutável.
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
            examples:
              expense:
                value:
                  type: EXPENSE
                  description: Mercado sintético
                  amount:
                    minor: 4250
                    currency: BRL
                  paymentMethod: PIX
                  categoryId: expense.food
                  occurredAt: "2026-08-14T15:00:00Z"
              income:
                value:
                  type: INCOME
                  description: Salário sintético
                  amount:
                    minor: 725000
                    currency: BRL
                  categoryId: income.salary
                  occurredAt: "2026-08-14T15:00:00Z"
      responses:
        "200":
          description: Representação canônica pronta para revisão
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPreview"
        "400":
          $ref: "#/components/responses/TransactionPreviewBadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/categories:
    get:
      operationId: listCategories
      summary: Lista o catálogo de categorias financeiras do sistema
      description: >-
        Retorna primeiro as categorias aplicáveis a Expense e depois as
        categorias aplicáveis a Income, em ordem estável do catálogo. A opção
        conceitual Sem categoria é representada pela ausência de categoryId em
        uma transaction e não integra esta lista.
      tags: [Categories]
      responses:
        "200":
          description: Catálogo completo e ordenado de categorias do sistema
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryCatalogResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/transactions:
    post:
      operationId: recordTransaction
      summary: Registra uma despesa ou receita já confirmada pelo canal
      description: >-
        Registra no organizador uma despesa ou receita já ocorrida e confirmada
        pelo usuário no canal chamador. Esta operação não executa pagamento,
        recebimento, Pix, transferência, compra nem movimentação bancária. O
        owner é o contexto single-owner temporário do servidor; origin IOS e
        timezone financeira America/Sao_Paulo também são atribuídos pelo
        servidor. Retry com a mesma chave, operação e semântica retorna o recurso
        original sem novas gravações.
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
            examples:
              expense:
                value:
                  type: EXPENSE
                  description: Mercado sintético
                  amount:
                    minor: 4250
                    currency: BRL
                  paymentMethod: PIX
                  categoryId: expense.food
                  occurredAt: "2026-08-14T15:00:00Z"
              income:
                value:
                  type: INCOME
                  description: Salário sintético
                  amount:
                    minor: 725000
                    currency: BRL
                  categoryId: income.salary
                  occurredAt: "2026-08-14T15:00:00Z"
      responses:
        "201":
          description: Movimentação registrada ou replay seguro do recurso original
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
            Idempotency-Replayed:
              description: Presente com valor true somente quando a resposta é replay.
              schema:
                type: string
                const: "true"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/TransactionCreateBadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listTransactionsByMonth
      summary: Lista despesas e receitas do owner no mês financeiro informado
      description: >-
        Usa limites de calendário em America/Sao_Paulo e consulta o intervalo
        [início do mês, início do próximo mês). Não calcula total, saldo,
        orçamento ou resumo.
      tags: [Transactions]
      parameters:
        - name: month
          in: query
          required: true
          description: Mês financeiro estrito no formato YYYY-MM.
          schema:
            type: string
            pattern: "^(?!0000-)[0-9]{4}-(0[1-9]|1[0-2])$"
          example: "2026-08"
      responses:
        "200":
          description: Lista mista determinística, vazia quando não há movimentações
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionMonth"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  headers:
    CacheControl:
      required: true
      description: Impede armazenamento da resposta.
      schema:
        type: string
        const: no-store
    ContentTypeOptions:
      required: true
      description: Impede MIME sniffing da resposta.
      schema:
        type: string
        const: nosniff
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Chave opaca de 1 a 128 bytes ASCII visíveis, escopada por owner e pela
        operação lógica. CREATE_EXPENSE, CREATE_INCOME, CREATE_RECURRENCE,
        CANCEL_RECURRENCE, CREATE_CREDIT_CARD, ARCHIVE_CREDIT_CARD,
        CREATE_CARD_PURCHASE e CANCEL_INSTALLMENT_PLAN possuem
        namespaces independentes. A mesma string pode
        ser usada em operações ou owners diferentes sem colisão. O valor nunca é
        devolvido em erros ou responses.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[!-~]+$"
      example: synthetic-create-20260814-001
...
  responses:
    InvalidRequest:
      description: Request inválido
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvalidRequestError"
    TransactionPreviewBadRequest:
      description: Request ou tipo de transaction inválidos
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransactionPreviewBadRequestError"
    TransactionCreateBadRequest:
      description: Request, tipo de transaction ou chave de idempotência inválidos
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransactionCreateBadRequestError"
...
    IdempotencyConflict:
      description: Chave reutilizada com semântica diferente
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IdempotencyConflictError"
          example:
            error:
              code: IDEMPOTENCY_KEY_REUSED
              message: idempotency key was reused with a different request
...
    MethodNotAllowed:
      description: Método não permitido para a rota financeira
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MethodNotAllowedError"
    InternalError:
      description: Falha interna sem exposição de detalhes técnicos
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalError"
  schemas:
...
    CategoryId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[a-z][a-z0-9]*([._][a-z0-9]+)*$"
      description: >-
        Identidade técnica estável de uma categoria do sistema. O campo é
        omitido para uma transaction sem categoria e nunca aceita null.
    ExpenseRequest:
      type: object
      additionalProperties: false
      required: [type, description, amount, paymentMethod, occurredAt]
      properties:
        type:
          type: string
          const: EXPENSE
        description:
          type: string
          minLength: 1
          maxLength: 200
        amount:
          $ref: "#/components/schemas/Money"
        paymentMethod:
          type: string
          enum: [PIX, DEBIT, CREDIT, CASH]
        categoryId:
          $ref: "#/components/schemas/CategoryId"
        occurredAt:
          type: string
          format: date-time
          description: >-
            Instante normalizado para UTC e truncado para precisão de
            microssegundos antes de preview, fingerprint, persistência e resposta.
    IncomeRequest:
      type: object
      additionalProperties: false
      required: [type, description, amount, occurredAt]
      properties:
        type:
          type: string
          const: INCOME
        description:
          type: string
          minLength: 1
          maxLength: 200
        amount:
          $ref: "#/components/schemas/Money"
        categoryId:
          $ref: "#/components/schemas/CategoryId"
        occurredAt:
          type: string
          format: date-time
          description: >-
            Instante normalizado para UTC e truncado para precisão de
            microssegundos antes de preview, fingerprint, persistência e resposta.
    TransactionRequest:
      oneOf:
        - $ref: "#/components/schemas/ExpenseRequest"
        - $ref: "#/components/schemas/IncomeRequest"
      discriminator:
        propertyName: type
        mapping:
          EXPENSE: "#/components/schemas/ExpenseRequest"
          INCOME: "#/components/schemas/IncomeRequest"
...
    TransactionPreview:
      oneOf:
        - $ref: "#/components/schemas/ExpensePreview"
        - $ref: "#/components/schemas/IncomePreview"
      discriminator:
        propertyName: type
        mapping:
          EXPENSE: "#/components/schemas/ExpensePreview"
          INCOME: "#/components/schemas/IncomePreview"
...
    TransactionResponse:
      oneOf:
        - $ref: "#/components/schemas/Expense"
        - $ref: "#/components/schemas/Income"
      discriminator:
        propertyName: type
        mapping:
          EXPENSE: "#/components/schemas/Expense"
          INCOME: "#/components/schemas/Income"
    TransactionMonth:
      type: object
      additionalProperties: false
      required: [month, items]
      properties:
        month:
          type: string
          pattern: "^(?!0000-)[0-9]{4}-(0[1-9]|1[0-2])$"
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponse"
...
    CategoryCatalogResponse:
      type: array
      description: >-
        Catálogo de sistema ordenado por aplicabilidade e ordem interna; não
        inclui uma entrada para Sem categoria.
      items:
        $ref: "#/components/schemas/CategoryDefinitionResponse"
...
    TransactionPreviewBadRequestError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNSUPPORTED_TRANSACTION_TYPE]
            message:
              type: string
...